requirements.txt
python-telegram-bot==20.0
pyresparser==0.6.4
transformers==4.21.1
jinja2==3.1.2
beautifulsoup4==4.10.0
spacy==3.2.4
pdfminer.six==20201018
fpdf==1.7.2
requests==2.26.0
torch==1.10.0
subprocess32==3.5.4